refactor(frontend): split home page components
Extract presentational components from +page.svelte:

- AppHeader.svelte: Header with mode toggle, Twitch status, actions
- AuthPanel.svelte: Unauthenticated UI (access code/QR login)
- YouTubeModeView.svelte: YouTube tabs wrapper
- TwitchChannelsView.svelte: Main Twitch channels view
- ChannelCard.svelte: Individual channel card
- AddChannelForm.svelte: Add channel form
- RecordingsOverview.svelte: Recordings management view
- ConfirmRemoveDialog.svelte: Remove confirmation modal

Also creates types.ts for component prop interfaces.

+page.svelte remains the state coordinator:
- All state management stays in the route
- Components receive props and callbacks
- No behavioral changes, no API changes

CSS Fixes:
- ChannelCard.svelte: Added border:0 and cursor:pointer to .watch-btn
- TwitchChannelsView.svelte: Added missing button properties to .add-btn
  (border-radius, font, font-weight, line-height, cursor, display, align-items,
   justify-content, min-height)
- +page.svelte: Added mobile media query for .panel padding

Additional Bug Fixes:
1. Version Line: Changed from position:fixed to position:absolute
   - Moved element inside .shell
   - Added position:relative to .shell
   - Version now scrolls with content

2. Live-Only Toggle: Fixed two-way binding
   - Changed from checked={liveOnly} to bind:checked={liveOnly}
   - Updated callback signature to pass value
   - Toggle now correctly updates state

3. YouTube Player Layout: Fixed loading state dimensions
   - Added .loading-box class with aspect-ratio:16/9
   - Added consistent sizing for error-box
   - Player maintains dimensions during load/empty/error states

Validation:
- pnpm run check: passed (0 errors, 0 warnings)
- pnpm run build: passed successfully
